Migrate to NextJS:   - Add SSR with nextjs   - Change structure to better align with nextjs   - Remove react router from components   - Add nextjs router/links
Start redoing profile component since migrating to
nextjs broke the nested router/navigation.
will add all components directly on profile page instead.
