+++ components/Library.js
@@ -5,8 +5,8 @@ import { Link } from "react-router-dom";
 import axios from "axios";
 import Paper from "@material-ui/core/Paper";
 import { makeStyles } from "@material-ui/core/styles";
-import UserContext from "../../context/UserContext";
-import Note from "../Note";
+import UserContext from "../context/UserContext";
+import Note from "./Note";
 
 const useStyles = makeStyles(() => ({
   paper: {
