+++ components/PostButtons.js
@@ -4,7 +4,7 @@
 import React, { useContext, useState, useEffect } from "react";
 import axios from "axios";
 import { useHistory } from "react-router-dom";
-import UserContext from "../../context/UserContext";
+import UserContext from "../context/UserContext";
 
 const PostButtons = (props) => {
   const { postData } = props;
