+++ components/UserProfileView.js
@@ -2,12 +2,12 @@
 /* eslint-disable no-underscore-dangle */
 import React, { useEffect, useState, useContext } from "react";
 import { makeStyles } from "@material-ui/core/styles";
-import { useParams } from "react-router-dom";
+import { useRouter } from "next/router";
 import Paper from "@material-ui/core/Paper";
 
 import axios from "axios";
-import UserContext from "../../context/UserContext";
-import profile from "../../media/profile.jpg";
+import UserContext from "../context/UserContext";
+import profile from "../public/profile.jpg";
 
 const useStyles = makeStyles(() => ({
   paper: {
@@ -41,22 +41,23 @@ const useStyles = makeStyles(() => ({
   },
 }));
 
-const UserProfileView = () => {
+const UserProfileView = (props) => {
+  const { userId } = props;
   const classes = useStyles();
+  const router = useRouter();
   const [user, setUser] = useState({});
-  const { id } = useParams();
 
   useEffect(() => {
     const getUser = async () => {
       try {
-        const fetchedUser = await axios.get(`http://localhost:5000/user/${id}`);
+        const fetchedUser = await axios.get(`http://localhost:5000/user/${userId}`);
         setUser(fetchedUser.data);
       } catch (error) {
         console.log(`THIS MESSAGE:${error}`);
       }
     };
     getUser();
-  }, [id]);
+  }, [userId]);
 
   return (
     <Paper elevation={3} className={classes.paper}>
