+++ context/UserContext.js
@@ -1,0 +1,8 @@
+import { createContext } from "react";
+
+export default createContext({
+  userData: {
+    token: null,
+    user: {},
+  }
+});
