+++ next.config.js
@@ -1,0 +1,3 @@
+const withImages = require("next-images");
+
+module.exports = withImages();
