+++ package.json
@@ -1,89 +1,25 @@
 {
-  "name": "u11-fullstack",
+  "name": "u11-fullstack-nextjs",
   "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "axios": "^0.21.1",
-    "bootstrap": "^4.5.3",
-    "eslint": "^7.10.0",
-    "express": "^4.17.1",
-    "next": "^10.0.5",
-    "prop-types": "^15.7.2",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-hook-form": "^6.9.2",
-    "react-router": "^5.2.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
-    "react-swipeable-views": "^0.13.9"
-  },
   "scripts": {
-    "build next": "next dev -p 1234",
-    "dev next": "next build",
-    "start next": "next start -p 1234",
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "lint": "eslint ."
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "airbnb",
-      "airbnb/hooks",
-      "plugin:jsx-a11y/recommended",
-      "prettier",
-      "prettier/react"
-    ],
-    "plugins": [
-      "jsx-a11y",
-      "prettier"
-    ],
-    "rules": {
-      "semi": 1,
-      "react/jsx-filename-extension": [
-        1,
-        {
-          "extensions": [
-            ".js",
-            ".jsx"
-          ]
-        }
-      ],
-      "prettier/prettier": [
-        "error",
-        {
-          "semi": true
-        }
-      ]
-    }
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
-  },
-  "devDependencies": {
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-prettier": "^6.12.0",
-    "eslint-plugin-import": "^2.21.2",
-    "eslint-plugin-jsx-a11y": "^6.3.0",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.20.0",
-    "eslint-plugin-react-hooks": "^4.0.0",
-    "prettier": "^2.1.2"
+  "dependencies": {
+    "@material-ui/core": "^4.11.2",
+    "@material-ui/icons": "^4.11.2",
+    "@svgr/webpack": "^5.5.0",
+    "@zeit/next-css": "^1.0.1",
+    "axios": "^0.21.1",
+    "next": "10.0.5",
+    "next-compose-plugins": "^2.2.1",
+    "next-images": "^1.6.2",
+    "next-react-svg": "^1.1.2",
+    "react": "17.0.1",
+    "react-dom": "17.0.1",
+    "react-hook-form": "^6.14.0",
+    "react-router-dom": "^5.2.0"
   }
 }
