+++ pages/login.jsx
@@ -4,22 +4,21 @@ import CssBaseline from "@material-ui/core/CssBaseline";
 import TextField from "@material-ui/core/TextField";
 import FormControlLabel from "@material-ui/core/FormControlLabel";
 import Checkbox from "@material-ui/core/Checkbox";
-import Link from "@material-ui/core/Link";
+import Link from "next/link";
 import Typography from "@material-ui/core/Typography";
 import { makeStyles } from "@material-ui/core/styles";
 import { useForm, Controller } from "react-hook-form";
 import axios from "axios";
-import { useHistory } from "react-router-dom";
+import { useRouter } from "next/router";
 import Paper from "@material-ui/core/Paper";
+import Header from "../components/Header";
 import UserContext from "../context/UserContext";
-import "../App.css";
 
 const useStyles = makeStyles((theme) => ({
   paper: {
     width: "100%",
     padding: "20px",
     borderRadius: "0px",
-    backgroundColor: "#9E8FB2",
     display: "flex",
     justifyContent: "center",
     alignItems: "center",
@@ -46,12 +45,13 @@ const useStyles = makeStyles((theme) => ({
   inputField: {
     width: "48%",
     minWidth: "150px",
+    margin: "10px",
   },
 }));
 
 export default function SignIn() {
   const classes = useStyles();
-  const history = useHistory();
+  const router = useRouter();
   const { register, handleSubmit, control } = useForm();
   const { setUserData } = useContext(UserContext);
 
@@ -69,20 +69,19 @@ export default function SignIn() {
       });
       localStorage.setItem("auth-token", loginRes.data.token);
       if (loginRes.data.user.role === "admin") {
-        history.push("/admin");
+        router.push("/admin");
       } else {
-        history.push(`/profile/${loginRes.data.user.id}`);
+        router.push(`/profile/${loginRes.data.user.id}`);
       }
     } catch (error) {
       console.log(`THIS MESSAGE:${error}`);
-      console.log(error.response.data);
-      console.log(error.response.status);
-      console.log(error.response.headers);
+      console.log(error);
+  
     }
   };
 
   return (
-    <Paper elevation={0} className={classes.paper}>
+    <Paper elevation={0} className={`alignHeader ${classes.paper}`}>
       <CssBaseline />
       <Typography component="h1" variant="h5">
         Sign in
@@ -141,14 +140,13 @@ export default function SignIn() {
         >
           Sign In
         </Button>
-
         <Link href="/register" variant="body2">
           Forgot password?
-        </Link>
+          </Link>
         <Link href="/register" variant="body2">
           Dont have an account? Sign Up
         </Link>
       </form>
     </Paper>
-  );
-}
+  )
+};
