--- src/index.jsx
@@ -1,20 +1,0 @@
-import React from "react";
-// eslint-disable-next-line no-unused-vars
-import { BrowserRouter, Route, Switch } from "react-router-dom";
-import CssBaseline from "@material-ui/core/CssBaseline";
-import ReactDOM from "react-dom";
-import App from "./App";
-import * as serviceWorker from "./serviceWorker";
-
-ReactDOM.render(
-  <BrowserRouter>
-    <CssBaseline />
-    <App />
-  </BrowserRouter>,
-  document.getElementById("root")
-);
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
