+++ styles/server.js
@@ -1,0 +1,35 @@
+import express from 'express';
+import React from 'react';
+import ReactDOMServer from 'react-dom/server';
+import { ServerStyleSheets, ThemeProvider } from '@material-ui/core/styles';
+import App from './App';
+import theme from './theme';
+
+function handleRender(req, res) {
+  const sheets = new ServerStyleSheets();
+
+  // Render the component to a string.
+  const html = ReactDOMServer.renderToString(
+    sheets.collect(
+      <ThemeProvider theme={theme}>
+        <App />
+      </ThemeProvider>,
+    ),
+  );
+
+  // Grab the CSS from the sheets.
+  const css = sheets.toString();
+
+  // Send the rendered page back to the client.
+  res.send(renderFullPage(html, css));
+}
+
+const app = express();
+
+app.use('/build', express.static('build'));
+
+// This is fired every time the server-side receives a request.
+app.use(handleRender);
+
+const port = 3000;
+app.listen(port);
